refactor: 구글로그인 버튼 삭제
