+++ package.json
@@ -22,7 +22,7 @@
     "react-intersection-observer": "^9.16.0",
     "react-toastify": "^11.0.5",
     "winedoojan": "file:",
-    "zod": "^3.24.3",
+    "zod": "^3.24.4",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
