+++ src/app/myprofile/page.tsx
@@ -1,20 +1,20 @@
-"use client";
+'use client';
 
-import MyReviewCard from "@/components/MyReviewCard";
-import { fetchReviews, fetchWines } from "../../services/myProfileApi";
-import { Wine, Review } from "@/types/myprofileTypes";
-import MyProfile from "@/components/MyProfile";
-import { useEffect, useState } from "react";
-import MyWineCard from "@/components/MyWineCard";
-import images from "../../../public/images/images";
-import BlobButton from "@/components/common/BlobButton";
-import Header from "@/components/common/Header";
-import { useAuthStore } from "@/store/authStore";
-import { useRouter } from "next/navigation";
-import { useInView } from "react-intersection-observer";
-import WineModal from "@/components/modals/WineModal";
-import Image from "next/image";
-import clsx from "clsx";
+import MyReviewCard from '@/components/MyReviewCard';
+import { fetchReviews, fetchWines } from '../../services/myProfileApi';
+import { Wine, Review } from '@/types/schema';
+import MyProfile from '@/components/MyProfile';
+import { useEffect, useState } from 'react';
+import MyWineCard from '@/components/MyWineCard';
+import images from '../../../public/images/images';
+import BlobButton from '@/components/common/BlobButton';
+import Header from '@/components/common/Header';
+import { useAuthStore } from '@/store/authStore';
+import { useRouter } from 'next/navigation';
+import { useInView } from 'react-intersection-observer';
+import WineModal from '@/components/modals/WineModal';
+import Image from 'next/image';
+import clsx from 'clsx';
 
 const MyCardSkeleton = () => (
   <div className="w-[343px] lg:w-[800px] md:w-[704px] flex flex-row gap-[16px] rounded-xl border border-gray-200 bg-white p-12 shadow-sm animate-pulse mb-2.5">
@@ -32,7 +32,7 @@ export default function ProfilePage() {
   const router = useRouter();
   const user = useAuthStore((state) => state.user);
   const accessToken = useAuthStore((state) => state.accessToken);
-  const [tab, setTab] = useState<"reviews" | "wines">("reviews");
+  const [tab, setTab] = useState<'reviews' | 'wines'>('reviews');
   const [openId, setOpenId] = useState<number | null>(null); // 햄버거 버튼 열기/닫기를 위한 상태
   const [myProfileData, setMyProfileData] = useState({
     reviews: [] as Review[],
@@ -48,17 +48,17 @@ export default function ProfilePage() {
   const limit = 3;
 
   // 리뷰/와인 목록 가져오기
-  const getList = async (tab: "reviews" | "wines", isLoadMore = false) => {
+  const getList = async (tab: 'reviews' | 'wines', isLoadMore = false) => {
     if (!user || !accessToken) return;
 
-    const cursorKey = tab === "reviews" ? "reviewCursor" : "wineCursor";
+    const cursorKey = tab === 'reviews' ? 'reviewCursor' : 'wineCursor';
     const cursor = isLoadMore ? myProfileData[cursorKey] : undefined;
     const listKey = tab;
 
     if (isLoadMore && cursor === null) return;
 
     const res =
-      tab === "reviews"
+      tab === 'reviews'
         ? await fetchReviews({
             teamId: user.teamId,
             limit,
@@ -80,7 +80,7 @@ export default function ProfilePage() {
     }));
   };
 
-  const loadData = async (tab: "reviews" | "wines", isLoadMore = false) => {
+  const loadData = async (tab: 'reviews' | 'wines', isLoadMore = false) => {
     // skeleton ui 보여주기
     if (!isLoadMore) setIsLoading(true);
     await getList(tab, isLoadMore);
@@ -114,7 +114,7 @@ export default function ProfilePage() {
   // 등록된 리뷰 없을 때 버튼 클릭 함수
   useEffect(() => {
     if (isReviewModalOpen) {
-      router.push("/wines");
+      router.push('/wines');
     }
   }, [isReviewModalOpen]);
 
@@ -149,8 +149,8 @@ export default function ProfilePage() {
 
       <div
         className={clsx(
-          "flex justify-start gap-[30px] md:gap-10 lg:gap-[60px] flex-col mx-auto w-[343px] mt-5 lg:mt-[37px]",
-          "lg:flex-row lg:w-[1140px] md:w-[704px]"
+          'flex justify-start gap-[30px] md:gap-10 lg:gap-[60px] flex-col mx-auto w-[343px] mt-5 lg:mt-[37px]',
+          'lg:flex-row lg:w-[1140px] md:w-[704px]'
         )}
       >
         {/* 프로필 이미지, 닉네임*/}
@@ -160,21 +160,21 @@ export default function ProfilePage() {
           {/* 탭 */}
           <div
             className={`flex gap-[32px] items-center ${
-              tab === "reviews" ? "mb-[22px]" : "mb-[64px]"
+              tab === 'reviews' ? 'mb-[22px]' : 'mb-[64px]'
             }`}
           >
             <button
-              onClick={() => setTab("reviews")}
+              onClick={() => setTab('reviews')}
               className={`w-max h-[32px] font-bold text-lg md:text-xl  ${
-                tab === "reviews" ? "text-[#2D3034]" : "text-[#9FACBD]"
+                tab === 'reviews' ? 'text-[#2D3034]' : 'text-[#9FACBD]'
               }`}
             >
               내가 쓴 후기
             </button>
             <button
-              onClick={() => setTab("wines")}
+              onClick={() => setTab('wines')}
               className={`w-max h-[32px] font-bold text-lg md:text-xl ${
-                tab === "wines" ? "text-[#2D3034]" : "text-[#9FACBD]"
+                tab === 'wines' ? 'text-[#2D3034]' : 'text-[#9FACBD]'
               }`}
             >
               내가 등록한 와인
@@ -194,8 +194,8 @@ export default function ProfilePage() {
                 <MyCardSkeleton />
                 <MyCardSkeleton />
               </>
-            ) : (tab === "reviews" && myProfileData.reviews.length === 0) ||
-              (tab === "wines" && myProfileData.wines.length === 0) ? (
+            ) : (tab === 'reviews' && myProfileData.reviews.length === 0) ||
+              (tab === 'wines' && myProfileData.wines.length === 0) ? (
               // {/* 데이터 없을 때 : 이미지 + 버튼 */}
               <div className="lg:w-[800px] lg:h-[530px] flex flex-col gap-[30px] items-center justify-center">
                 <img
@@ -204,14 +204,14 @@ export default function ProfilePage() {
                   className="size-[180px]"
                 />
                 <div className="font-bold text-2xl text-[#2D3034]">
-                  등록된 {tab === "reviews" ? "리뷰가" : "와인이"} 없어요
+                  등록된 {tab === 'reviews' ? '리뷰가' : '와인이'} 없어요
                 </div>
                 <BlobButton
                   children={
-                    tab === "reviews" ? "리뷰등록하러가기" : "와인등록하러가기"
+                    tab === 'reviews' ? '리뷰등록하러가기' : '와인등록하러가기'
                   }
                   onClick={() => {
-                    if (tab === "reviews") {
+                    if (tab === 'reviews') {
                       setIsReviewModalOpen(true);
                     } else {
                       setIsWineModalOpen(true);
@@ -222,7 +222,7 @@ export default function ProfilePage() {
             ) : (
               //{/* 데이터 있을 때 : 와인/리뷰 카드 */}
               <>
-                {tab === "reviews"
+                {tab === 'reviews'
                   ? myProfileData.reviews.map((review) => (
                       <MyReviewCard
                         key={review.id}
